# docs/v2/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Neighbors REST API",
    "version": "2.0",
    "license": {
      "name": "Gnu General Public License",
      "url": "https://www.gnu.org/licenses/gpl.html"
    },
    "description": "The Neighbors REST API gives access to genome sequences via the taxonomy of sequenced life.\nStarting from one or more target organisms, the accessions of their genomes can be looked up.\nIn addition, the genome accessions of the targets' closest relatives, their neighbors, can be looked up.\nSuch pairs of samples of target and neighbor genomes make good starting material for developing diagnostic markers.",
    "contact": {
      "url": "https://www.evolbio.mpg.de/16023/group_bioinformatics",
      "name": "Bernhard Haubold",
      "email": "haubold@evolbio.mpg.de"
    }
  },
  "servers": [
    {
      "url": "https://neighbors.evolbio.mpg.de/api/v2"
    }
  ],
...
  "paths": {
    "/accessions": {
      "get": {
        "summary": "Get accession levels",
        "tags": [
          "Accession"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Accession"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "accession": "GCF_000001405.40",
                        "level": "chromosome"
                      },
                      {
                        "accession": "GCA_000002115.2",
                        "level": "chromosome"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter."
          }
        },
        "operationId": "get-accessions",
        "parameters": [
          {
            "schema": {
              "type": "array",
              "example": [
                "GCF_000001405.40",
                "GCA_000002115.2"
              ],
              "items": {
                "type": "string"
              }
            },
            "in": "query",
            "name": "accession_ids",
            "description": "IDs of accessions to filter for",
            "explode": false,
            "style": "form",
            "required": true
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/page_size"
          }
        ],
        "description": "Get accessions and their levels filtered by accession IDs"
      }
    },
...
    "/ranks": {
      "get": {
        "summary": "Get ranks of taxa",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rank"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "tax_id": 9606,
                        "rank": "species"
                      },
                      {
                        "tax_id": 9605,
                        "rank": "genus"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter."
          }
        },
        "operationId": "get-ranks",
        "parameters": [
          {
            "schema": {
              "type": "array",
              "example": [
                9606,
                9605
              ],
              "items": {
                "type": "integer"
              }
            },
            "in": "query",
            "name": "taxon_ids",
            "description": "IDs of taxa to filter for",
            "explode": false,
            "style": "form",
            "required": true
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/page_size"
          }
        ],
        "description": "Get ranks of taxa filtered by taxon IDs"
      }
    },
    "/taxa": {
      "get": {
        "summary": "Get taxa by name",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Taxon"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "taxid": 103600,
                        "parent": 103599,
                        "name": "Sousa chinensis",
                        "common_name": "Indo-pacific humpbacked dolphin"
                      },
                      {
                        "taxid": 9737,
                        "parent": 9734,
                        "name": "Stenella coeruleoalba",
                        "common_name": "striped dolphin"
                      },
                      {
                        "taxid": 79784,
                        "parent": 9738,
                        "name": "Tursiops aduncus",
                        "common_name": "Indo-pacific bottlenose dolphin"
                      },
                      {
                        "taxid": 34814,
                        "parent": 13453,
                        "name": "Coryphaena hippurus",
                        "common_name": "common dolphinfish"
                      },
                      {
                        "taxid": 27766,
                        "parent": 1489907,
                        "name": "Coryphaenidae",
                        "common_name": "dolphinfishes"
                      },
                      {
                        "taxid": 118797,
                        "parent": 118796,
                        "name": "Lipotes vexillifer",
                        "common_name": "Yangtze River dolphin"
                      },
                      {
                        "taxid": 48752,
                        "parent": 48750,
                        "name": "Platanista minor",
                        "common_name": "Indus River dolphin"
                      },
                      {
                        "taxid": 3371155,
                        "parent": 3371042,
                        "name": "Sagmatias obliquidens",
                        "common_name": "Pacific white-sided dolphin"
                      },
                      {
                        "taxid": 46167,
                        "parent": 46166,
                        "name": "Steno bredanensis",
                        "common_name": "rough-toothed dolphin"
                      },
                      {
                        "taxid": 103582,
                        "parent": 27607,
                        "name": "Cephalorhynchus commersonii",
                        "common_name": "Commerson's dolphin"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter."
          }
        },
        "operationId": "get-taxa",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "dolph"
            },
            "in": "query",
            "name": "name",
            "description": "Name or pattern of the taxa",
            "required": true
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "scientific",
            "description": "Whether to search in scientific names only"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "exact",
            "description": "Whether to search for exactly that name or use the name as a pattern"
          },
          {
            "schema": {
              "type": "integer",
              "example": 2,
              "default": 0,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "description": "Current page"
          },
          {
            "schema": {
              "type": "integer",
              "example": 10,
              "default": -1,
              "minimum": -1
            },
            "in": "query",
            "name": "page_size",
            "description": "Amount of entries per page"
          }
        ],
        "description": "Get basic taxon data for taxa with the given name"
      }
    },
    "/taxa-accessions": {
      "get": {
        "summary": "Get taxons accessions",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonAccessions"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "taxid": 278148,
                        "accessions": [
                          {
                            "accession": "GCF_001618845.1",
                            "level": "complete"
                          }
                        ]
                      },
                      {
                        "taxid": 602633,
                        "accessions": [
                          {
                            "accession": "GCA_003063835.1",
                            "level": "scaffold"
                          }
                        ]
                      },
                      {
                        "taxid": 765698,
                        "accessions": [
                          {
                            "accession": "GCF_000185905.1",
                            "level": "complete"
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter."
          }
        },
        "operationId": "get-taxa-accessions",
        "description": "Get the accessions associated with taxa filtered by taxon IDs",
        "parameters": [
          {
            "schema": {
              "type": "array",
              "example": [
                278148,
                602633
              ],
              "items": {
                "type": "integer"
              }
            },
            "in": "query",
            "name": "taxon_ids",
            "description": "IDs of taxa to filter for",
            "explode": false,
            "style": "form",
            "required": true
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/page_size"
          }
        ]
      }
    },
...
    "/taxa-names": {
      "get": {
        "summary": "Get taxa and their names",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonName"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "tax_id": 9606,
                        "name": "Homo sapiens",
                        "common_name": "human"
                      },
                      {
                        "tax_id": 9605,
                        "name": "Homo",
                        "common_name": ""
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter."
          }
        },
        "operationId": "get-taxa-names",
        "description": "Get taxa and their scientific and common names filtered by taxon IDs",
        "parameters": [
          {
            "schema": {
              "type": "array",
              "example": [
                9606,
                9605
              ],
              "items": {
                "type": "integer"
              }
            },
            "in": "query",
            "name": "taxon_ids",
            "description": "IDs of taxa to filter for",
            "explode": false,
            "style": "form",
            "required": true
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/page_size"
          }
        ]
      }
    },
...
    "/taxa/{taxon_id}/children": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "example": 9606
          },
          "name": "taxon_id",
          "in": "path",
          "required": true,
          "description": "id of taxon"
        }
      ],
      "get": {
        "summary": "Get taxons children",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonName"
                  }
                },
                "examples": {
                  "Example": {
                    "value": [
                      {
                        "taxid": 741158,
                        "name": "Homo sapiens subsp. 'Denisova'",
                        "common_name": "Denisova hominin"
                      },
                      {
                        "taxid": 63221,
                        "name": "Homo sapiens neanderthalensis",
                        "common_name": "Neandertal"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "operationId": "get-taxa-taxon-id",
        "description": "Get the children of a taxon given its id",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/page_size"
          }
        ]
      }
    },
...
    "/taxa/{taxon_id}/parent": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "example": 9606
          },
          "name": "taxon_id",
          "in": "path",
          "required": true,
          "description": "id of taxon"
        }
      ],
      "get": {
        "summary": "Get taxons parent",
        "tags": [
          "Taxon"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxId"
                },
                "examples": {
                  "Example": {
                    "value": {
                      "tax_id": 9605
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-taxa-taxon-id-parent",
        "description": "Get the parents taxon id of a given taxon"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Accession": {
        "title": "Accession",
        "type": "object",
        "properties": {
          "accession": {
            "type": "string"
          },
          "level": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {},
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0,
          "example": 0,
          "minimum": 0
        },
        "description": "Current page",
        "style": "form"
      },
      "page_size": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": -1,
          "example": 20,
          "default": -1
        },
        "description": "Amount of entries per page",
        "style": "form"
      }
    }
  }
}